+++ semver-project-gradle-plugin/main/kotlin/com/javiersc/semver/project/gradle/plugin/tasks/PrintSemverTask.kt
@@ -1,6 +1,7 @@
 package com.javiersc.semver.project.gradle.plugin.tasks
 
 import com.javiersc.gradle.project.extensions.isRootProject
+import com.javiersc.kotlin.stdlib.or
 import com.javiersc.semver.project.gradle.plugin.internal.logOnlyOnRootProject
 import com.javiersc.semver.project.gradle.plugin.internal.projectTagPrefix
 import com.javiersc.semver.project.gradle.plugin.internal.semverMessage
@@ -99,7 +100,7 @@ constructor(
 
     @TaskAction
     public fun run() {
-        versions.orNull?.checkVersionIsHigherOrSame()
+        versions.orNull?.checkVersionIsHigherOrSame(version.orNull)
 
         val shouldLogMessage: Boolean = shouldLogMessage.get()
         val semver: String = version.get()