+++ semver-project-gradle-plugin/main/kotlin/com/javiersc/semver/project/gradle/plugin/tasks/PushSemverTagTask.kt
@@ -31,7 +31,7 @@ public abstract class PushSemverTagTask : DefaultTask() {
 
     @TaskAction
     public fun run() {
-        versions.orNull?.checkVersionIsHigherOrSame()
+        versions.orNull?.checkVersionIsHigherOrSame(version.orNull)
 
         gitTagBuildService
             .get()