+++ semver-project-gradle-plugin/main/kotlin/com/javiersc/semver/project/gradle/plugin/tasks/WriteSemverTask.kt
@@ -47,7 +47,7 @@ constructor(
 
     @TaskAction
     public fun run() {
-        versions.orNull?.checkVersionIsHigherOrSame()
+        versions.orNull?.checkVersionIsHigherOrSame(version.orNull)
 
         val semver: String = version.get()
         val prefix: String = tagPrefix.get()